split graphite/client.go for readability
